backend/requirements.txt
# ── Core Framework ──────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.0

# ── Document Processing ────────────────────────────────
python-docx>=1.1
pywin32>=306

# ── LLM ────────────────────────────────────────────────
openai>=1.30
httpx>=0.27

# ── Embeddings & Reranking ─────────────────────────────
sentence-transformers>=3.0
torch>=2.2
numpy<2

# ── PostgreSQL (async) ─────────────────────────────────
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9

# ── Qdrant ─────────────────────────────────────────────
qdrant-client>=1.9

# ── Redis ──────────────────────────────────────────────
redis[hiredis]>=5.0

# ── Web Search (Ninh Bình info) ─────────────────────────
wikipedia>=1.4.0
